#451 Putting the tests at the right place
